% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsAligment.R
\name{alignDNASubread}
\alias{alignDNASubread}
\title{Local alignment of short DNA-Seq reads by seed-and-vote algorithm}
\usage{
alignDNASubread(
  genome,
  fastqDir,
  fastq1,
  fastq2 = NULL,
  bamDir,
  bamFile,
  orientation = "fr",
  threads = 1,
  SV = FALSE,
  workDir = NULL
)
}
\arguments{
\item{genome}{character vector giving the path to and basename of index file.}

\item{fastqDir}{character vector giving the path to and name of directory
with FASTQ file(-s).}

\item{fastq1}{character vector including the name of FASTQ file containing
DNA-Seq reads to be aligned. For paired-end reads, this gives the name
of FASTQ file for first reads in each DNA-Seq library.}

\item{fastq2}{character vector including the name of FASTQ file containing
second DNA-Seq reads to be aligned. NULL by default.}

\item{bamDir}{character vector giving the path to and name of output
directory for BAM file.}

\item{bamFile}{character vector giving the name of output BAM file.}

\item{orientation}{character string giving the orientation of the two reads
from the same pair. Default value is "fr" (forward for the first read
and reverse for second one).}

\item{threads}{numeric value giving the number of threads used for mapping.
Default value is 1.}

\item{SV}{logical value indicating if structural variants will be detected
during read mapping. Default value is FALSE.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
BAM and VCF (if SV argument is TRUE) file(-s) are written to the
bamDir directory.
}
\description{
This is a wrapper function that uses the low-level function
align() from package Rsubread to align the short DNA-Seq reads by
seed-and-vote algorithm.
}
\examples{
genome <- "Reference_Genomes/GRCh38/GRCh38"
fastq1 <- c("example_seq.read1.fastq.gz")
fastq2 <- c("example_seq.read1.fastq.gz")
fastqDir <- "Files_FASTQ"
bamDir <- "Files_BAM"
bam <- "example_seq.bam"
workDir <- "/mnt/data/grinev"
al <- alignDNASubread(genome=genome,
                      fastqDir=fastqDir, fastq1=fastq1, fastq2=fastq2,
                      bamDir=bamDir,
                      bamFile=bam,
                      orientation="fr",
                      threads=28,
                      SV=FALSE,
                      workDir=workDir)
}
\author{
Vasily V. Grinev
}
