% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatingAndMeasuringEffects.r
\name{annotateDbSNP}
\alias{annotateDbSNP}
\title{Annotation of single nucleotide variations by dbSNP}
\usage{
annotateDbSNP(
  annoDir = NULL,
  annoFile,
  output = NULL,
  dbDir = NULL,
  dbFile = "GCF_000001405.40.gz",
  genome = "hg38",
  chrToNC = FALSE,
  chrID,
  workDir = NULL
)
}
\arguments{
\item{annoDir}{character string giving the name (or path to and name) of
directory with TXT file in tab-delimited format with annotated single nucleotide
variations. annoDir is NULL by default, which means the current
working directory will be used instead of the specified annotation directory.}

\item{annoFile}{character string giving the name of input TXT tab-delimited file.
By default, using of annotateBasic() output file is assumed. File must contains next columns:
i) seqnames - name of chromosome or scaffold;
ii) start - start coordinate of variation;
iii) end - end coordinate of variation;
iv) strand - strand of DNA;
iv) ref - reference allele;
v) alt - alt allele, called in this position.
All extra columns wont be used in comparison and will be added in output
file without changes.}

\item{output}{character string giving the name (without extension) of TXT
file in tab-delimited format for storing annotation. The postfix "dbSNP annotation"
will be added to this name. NULL by default, which means that the name will be generated
automatically based on the name of the input VCF file.}

\item{dbDir}{character string giving the path to directory with VCF file of dbSNP release.
NULL by default, which means the current working directory will be used instead of
the specified directory. Can be same with annoDir.}

\item{dbFile}{character string giving the name of dbSNP VCF file for
extended annotation. "GCF_000001405.40.gz" by default.}

\item{genome}{character string giving the version of genome. "hg38" by default. Used
for annotaded SNV subsetting from dbSNP.}

\item{chrToNC}{logical. If TRUE, sequence names in basic annotation will be
converted from UCSC-style format (chr) to NCBI-style (NC_***). Requires
an file with possible sequence names, provided in chrID. At the end of annotation,
UCSC format of names will be automatically restored. FALSE by default.}

\item{chrID}{character string giving the name tab-delimited TXT file with names
of chrmosome sequences in NCBI-RefSeq format (first column) and UCSC format
(second column). File must be placed into dbDir.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
an tab-delimited TXT file with basic SNVs annotation and additional fields
for variations that were found into dbSNP (for others columns will contain NA).
}
\description{
This function compares basic annotation of variants, produced by
annotateBasic() function, with annotation from dbSNVs current release.
Current version of function was successfully tested on GCF_000001405.40 release of
dbSNP. List of extracted dbSNP fields is fixed for now and cannot be
changed.
}
\examples{
annotateDbSNP(annoDir="annotation_dir",
             annoFile="SNVs, basic annotation.txt",
             dbDir="database_dir", dbFile="GCF_000001405.40.gz",
             chrToNC=TRUE, chrID="RefSeq Accessions & UCSC chromosomes.txt",
             workDir="D:/VCF_calling")
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
