% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsAligment.R
\name{assessQAlignedReads}
\alias{assessQAlignedReads}
\title{Evaluate alignment quality metrics}
\usage{
assessQAlignedReads(bamDir = NULL, bamFile, workDir = NULL)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character vector giving the name of BAM file to be evaluated.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
object of class list containing all main quality metrics for aligned
short reads.
}
\description{
Evaluation of the alignment quality metrics for aligned DAN-Seq
and/or RNA-Seq short reads.
}
\examples{
qal <- assessQAlignedReads(bamDir="Files_BAM",
                           bamFile="SRR21721515.bam",
                           workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
