% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawReadsQC.R
\name{assessQRawReads}
\alias{assessQRawReads}
\title{Calculate quality scores for short reads}
\usage{
assessQRawReads(
  fastqDir = NULL,
  fastq,
  n = NULL,
  adapters = NULL,
  contaminants = NULL,
  workDir = NULL
)
}
\arguments{
\item{fastqDir}{character string specifying the name of the directory
containing the FASTQ file(-s). NULL by default that mean the current
working directory.}

\item{fastq}{character string specifying the name of the FASTQ file
containing short reads. Allowed formats are "fastq" or "fastq.gz".}

\item{n}{integer value giving the number of reads loaded into the computer's
RAM. The default value is NULL, at which all reads are loaded into RAM.}

\item{adapters}{character string specifying the name of the tab-delimited
TXT file containing adapter sequences. The default value is NULL, at
which the search for adaptor sequences in short reads is not carried out.
If so, this file must contain the following three fields:
i) adapter_id         - adapter ID;
ii) adapter_name      - adapter name;
iii) adapter_sequence - adapter sequence.}

\item{contaminants}{character string specifying the name of the
tab-delimited TXT file containing foreign sequences. The default value
is NULL, at which the search for foreign sequences in short reads is not
carried out. If so, this file must contain the following three fields:
i) foreign_id         - foreign sequence ID;
ii) foreign_name      - foreign sequence name;
iii) foreign_sequence - foreign sequence itself.}

\item{workDir}{character string specifying the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
object of class list containing all main quality scores for short
reads.
}
\description{
Calculation of the quality scores for short reads.
}
\examples{
qa <- assessQRawReads(fastqDir="Files_FASTQ",
                      fastq="2-Galkina-A_S2_L001_R1_001.fastq.gz",
                      n=1e5,
                      adapters=NULL,
                      contaminants=NULL,
                      workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
