% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteringAndBenchmarkingVariants.r
\name{benchmarkVariants}
\alias{benchmarkVariants}
\title{Benchmark called variants against reference variants}
\usage{
benchmarkVariants(
  vcfDir,
  ref_vcf,
  query_vcf,
  featureTable = FALSE,
  workDir = NULL
)
}
\arguments{
\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(-s). NULL by default, which means the current
working directory will be used instead of the specified VCF directory.}

\item{ref_vcf}{character string giving the name of VCF file with reference
variants. Valid data format is *.vcf.}

\item{query_vcf}{character string giving the name of VCF file with called
variants. Valid data format is *.vcf.}

\item{featureTable}{logical argument, whether to create the feature table
that links each variant to its predicted classification outcome
(TP - true positive, FP - false positive, FN - false negative). Default
value is FALSE.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that mean the current working directory.}
}
\value{
consolidated data frame containing values of benchmarking metrics;
in addition, if featureTable=TRUE, in the working directory is writing
CSV file containing a detailed report of the benchmarking results.
}
\description{
This function performs benchmarking of called variants against
reference variants based on chromosome, position and allele identity.
}
\examples{
res <- benchmarkVariants(vcfDir="Files_VCF",
                         ref_vcf="test_truth.vcf",
                         query_vcf="test_query.vcf",
                         featureTable=FALSE,
                         workDir="D:/Vasily Grinev")
}
\author{
Liudmila S. Varaniuk.
}
