% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsAligment.R
\name{buildIndexSubread}
\alias{buildIndexSubread}
\title{Build a hash table for the reference genome}
\usage{
buildIndexSubread(
  ref_genome,
  ref_fasta,
  index,
  fa,
  memory = 8000,
  workDir = NULL
)
}
\arguments{
\item{ref_genome}{character string giving the name of folder for storing
of index for reference genome.}

\item{ref_fasta}{character string giving the name of folder for storing
of reference sequences.}

\item{index}{character string giving the basename of created index files.}

\item{fa}{charater string giving the name of the FASTA file containing all
the reference sequences.}

\item{memory}{a numeric value specifying the amount of memory (in megabytes)
used for storing the index during read mapping. Default value is 8000 MB.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
no value is produced but index files are written to the respective
directory.
}
\description{
This is a wrapper function that uses the low-level function
buildindex() from package Subread to build the hash table for the
reference genome.
}
\examples{
hash <- buildIndexSubread(ref_genome="Reference_Genomes",
                          ref_fasta="Files_FASTA",
                          index="GRCh38",
                          fa="GRCh38.fa",
                          memory=8000,
                          workDir="/mnt/data/grinev")
}
\author{
Vasily V. Grinev
}
