% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calling.R
\name{callSNVs}
\alias{callSNVs}
\title{Call single nucleotide variations using different statistical tests}
\usage{
callSNVs(
  fr_data,
  criteria = "entropy",
  vcfDir = NULL,
  outputVCF,
  bamDir = NULL,
  bamFile = NA,
  fastaDir = NULL,
  faFile = NA,
  baseq = 20,
  mindepth = 1,
  maxdepth = 1e+06,
  qvalue = 12,
  trim = 0,
  threads = 1,
  postfix = NA,
  intervals = NULL,
  pcr_model = NA,
  soft_clip = TRUE,
  ERC = "NONE",
  cores = "MAX",
  SE = TRUE,
  gatk_path = NULL,
  workDir = NULL,
  model = "rpart",
  pvalue = FALSE
)
}
\arguments{
\item{fr_data}{either a name of CSV file containing pileup statistics (as
returned by collectBasesPileup() function) or an object of class
data.frame with frequancy table.}

\item{criteria}{character vector specifying which test(-s) to be performed.
The following values are possible: "binomial", "counting", "entropy"
(by default), "fisher", "fisherSmyth", "gatk" and "poisson".}

\item{vcfDir}{character string giving the name (or path to and name) of
directory for storing of generated VCF file(-s). NULL by default, which
means the current working directory will be used instead of the
specified BAM directory.}

\item{outputVCF}{character string giving the name of output VCF file.}

\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(-s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{fastaDir}{character string giving the name (or path to and name) of
folder for storing of reference genome (with .fai indices if GATK).}

\item{faFile}{charater string giving the name of file containing reference
genome in format *.FASTA or *.FA.}

\item{baseq}{integer value, minimal "base quality" for each nucleotide in an
alignment. Default value is 20 that means read nucleotides with quality
scores less than 20 will be excluded from analysis.}

\item{mindepth}{integer value giving the minimal coverage of location
containing the variation. Default value is 1.}

\item{maxdepth}{integer value giving the maximal coverage of location
containing the variation. Default value is 1e6. This option is useful
for removing PCR artifacts.}

\item{qvalue}{numeric value giving the q-value cutoff for variations calling
at sequencing depth 50X. Default value is 12. The q-value is calcuated
as -log10(p), where p is a p-value yielded from the Fisher’s exact test.
The function exactSNP() automatically adjusts the q-value cutoff for
each chromosomal location according to its sequencing depth, based on
this cutoff.}

\item{trim}{integer value giving the number of nucleotides trimmed off from
each end of the read. Default value is 0.}

\item{threads}{numeric value giving the number of threads/CPUs used. Default
value is 1.}

\item{postfix}{character string that will be added to the output VCF file
name as postfix. NA by default at which the postfix "GATK based SNVs"
will be used. Only with GATK.}

\item{intervals}{character string specifying the name of the tab-delimited
TXT file containing coordinates of genomic location(-s) of interest. The
default value is NULL. If so, this file must contains the following four
fields:
i) seqnames - the name of chromosome or scaffold with prefix "chr";
ii) start   - start coordinate of genomic interval of interest;
iii) end    - end coordinate of genomic interval of interest;
iv) strand  - (optionally) strand information about genomic interval
of interest.
TXT file with genomic location(-s) must be in working directory.}

\item{pcr_model}{GATK argument --pcr-indel-model. Possible values: NONE,
AGRESSIVE, HOSTILE, CONSERVATIVE. See HaplotypeCaller
at https://gatk.broadinstitute.org for further details.}

\item{soft_clip}{GATK argument --dont-use-soft-clipped. TRUE by default.}

\item{ERC}{GATK ERC mode. "NONE" by default. Possible values: NONE, GVCF or
BP_RESOLUTION. See HaplotypeCaller at https://gatk.broadinstitute.org
for further details.}

\item{cores}{number of CPU cores to use for Hidden Markov Models. If "MAX"
(by default), all available cores will be detected and used.}

\item{SE}{if TRUE (by default), standart error determined by GATK will be
writen in file STDERR.HC.TXT.}

\item{gatk_path}{character string giving the path to internal or external
(system) GATK executive script. This path can be obtained by running
Find.Gatk() or settled manually. Default value is "gatk", i.e. function
will uses GATK systemly from $PATH.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
VCF file(-s) with detected single nucleotide variations.
}
\description{
This function is a wrapper that allows to call and control
various low-level functions for identification of single nucleotide
variations from NGS data.
}
\examples{
...
}
\author{
Dzianis D. Sarnatski, Mikalai M. Yatskou, Vasily V. Grinev
}
