% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBasesPileup.r
\name{collectBasesPileup}
\alias{collectBasesPileup}
\title{Calculate pileup statistics for a BAM file}
\usage{
collectBasesPileup(
  bamDir = NULL,
  bamFile,
  gr = "SeqInfo",
  canChr = TRUE,
  depth = 1e+06,
  baseq = 20,
  mapq = 20,
  coverage = NULL,
  genome = "BSgenome.Hsapiens.UCSC.hg38",
  tmpDir = NULL,
  postfix = NULL,
  workDir = NULL
)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{gr}{character string indicating the method for specifying the genomic
location(-s) for which pileup statistics will be collected. Default
value is "SeqInfo". In this case, the locations will be extracted
directly from the BAM file and pileup statistics will be collected for
the entire genome. Alternatively, the user can specify the genomic
locations of interest by applying to the argument a name of the
tab-delimited TXT file containing genomic coordinates. This file must
contains the following four fields:
i) seqnames - the name of chromosome or scaffold with prefix "chr";
ii) start   - start coordinate of genomic interval of interest;
iii) end    - end coordinate of genomic interval of interest;
iv) strand  - (optionally) strand information about genomic interval
of interest.
TXT file with genomic location(-s) must be in working directory.}

\item{canChr}{logical argument specifying that the records from non-canonical
chromosomes/scaffolds should be removed. Default value is TRUE.}

\item{depth}{integer value specifying the maximum number of overlapping
alignments considered for each position in the pileup.
Default value is 1e6.}

\item{baseq}{integer value specifying the minimal base quality for each
nucleotide in an alignment to be included in pileup. Default value is 20.}

\item{mapq}{integer value specifying the minimal mapping quality for an
alignment to be included in pileup. Default value is 20.}

\item{coverage}{integer value specifying the minimal coverage threshold for
filtering of successful genomic positions.}

\item{genome}{character string giving the name of the BSgenome data package
to be used for reference sequence extraction. Default value is
package BSgenome.Hsapiens.UCSC.hg38.}

\item{tmpDir}{character string giving the name (or path to and name) of
directory for storing of temporary file(-s) with pileup statistics.
NULL by default, which means the current working directory will be used
instead of the specified directory.}

\item{postfix}{character string that will be added to the CSV output file
name as postfix. NULL by default. In this case, the postfix
"PileupResults" will be used.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
a CSV file containing pileup statistics will be written to the
directory where input BAM file is stored.
}
\description{
This is a high-level function that performs a multi-threaded
calculation of the nucleotide-wise read coverage based on the BAM file
of interest.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{- The function processes the BAM file in parallel chunks, each
corresponding to a genomic range.
- It computes the number of occurrences of each nucleotide (A, C, G, T)
at each position within the specified range.
- The function includes a coverage threshold to filter positions with
low coverage.
- It uses reference genome to retrieve the reference nucleotide for each
position in the pileup data.
- The final pileup statistics is saved in the specified CSV file, which
contains information about the position, reference nucleotide, and
counts for A, C, G and T.
}\if{html}{\out{</div>}}
}
\examples{
res <- collectBasesPileup(bamDir="Files_BAM",
                          bamFile="SRR21721515.bam",
                          gr="genomic_intervals.txt",
                          canChr=TRUE,
                          depth=1e6,
                          baseq=20,
                          mapq=20,
                          coverage=10,
                          genome="BSgenome.Hsapiens.UCSC.hg38",
                          tmpDir=NULL,
                          postfix="PileupResults",
                          workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev, Dzianis D. Sarnatski, Mikalai M. Yatskou
}
