% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutRigthAdapter.r
\name{cutRigthAdapter}
\alias{cutRigthAdapter}
\title{Cut out adapter sequence(-s) from DNA- and/or RNA-Seq short reads}
\usage{
cutRigthAdapter(
  fastqDir = NULL,
  fastq,
  adapters,
  error = 0.2,
  min_match_flank = 3L,
  anchored = TRUE,
  indels = FALSE,
  workDir = NULL
)
}
\arguments{
\item{fastqDir}{character string giving the name (or path to and name) of
directory with FASTQ file(-s). NULL by default, which means the current
working directory will be used instead of the specified FASTQ directory.}

\item{fastq}{character string giving the name of FASTQ file containing short
reads of interest. Alternatively, fastq can be an internal R object with
preloaded FASTQ data.}

\item{adapters}{character string specifying the name of the tab-delimited
TXT file containing adapter sequences. This file must contain the
following three fields:
i) adapter_id         - adapter ID;
ii) adapter_name      - adapter name;
iii) adapter_sequence - adapter sequence.}

\item{error}{numeric value in \link{0, 1} specifying error rate. The error rate
is the proportion of mismatches allowed between the adapter and the
aligned portion of the read. For a given adapter A, the number of
allowed mismatches between each subsequence s of A and the read is
computed as: error * L_s, where L_s is the length of the subsequence s.
Default value is 0.2.}

\item{min_match_flank}{integer value giving the number of nucleotides.
Do not trim in flanks of the read if a match has min_match_flank of less
length. Default value is 3L (trim only with >=4 coincidences in a
flanking match).}

\item{anchored}{logical, TRUE by default. Can the adapter or partial adapter
be within (anchored=FALSE) or only in the terminal region(-s)
(anchored=TRUE) of the read?}

\item{indels}{logical, FALSE by default. If TRUE, indels are allowed in the
alignments of the suffixes of adapter with the read, at its beginning.}

\item{workDir}{character string specifying the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
internal R object containing short reads with trimmed adapter
sequence(-s).
}
\description{
Generic function for trimming of adapter sequence(-s) from
DNA- and/or RNA-Seq short reads. This function is based on low-level
functions cutRseq() of R/Bioconductor package FastqCleaner and
trimLRpattern() of R/Bioconductor package Biostrings.
}
\examples{
trimL <- cutRigthAdapter(fastqDir="Files_FASTQ",
                         fastq="example_seq.read1.fastq",
                         adapters="adapters.txt",
                         error=0.2,
                         min_match_flank=3L,
                         anchored=TRUE,
                         indels=FALSE, 
                         workDir="D:/Vasily Grinev")
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
