% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatingAndMeasuringEffects.r
\name{developAnnotations}
\alias{developAnnotations}
\title{Develop Ensembl-like annotations of experimental transcriptome}
\usage{
developAnnotations(
  gtfDir = NULL,
  gtfFile,
  orf,
  gtf,
  src,
  organism,
  sqlite,
  workDir = NULL
)
}
\arguments{
\item{gtfDir}{character string giving the name (or path to and name) of
directory with GTF file(-s). NULL by default, which means the current
working directory will be used instead of the specified GTF directory.}

\item{gtfFile}{character string giving the name of input GTF/GFF file with
experimental transcriptome. It is typically StringTie output. Valid
format is "gtf" or "gff".}

\item{orf}{character string giving the name of tab-delimited TXT file with
coordinates of open reading frames in experimental transcripts. It is
typically ORFhunteR output. This file should include three fields:
i) transcript_id - transcript ID;
ii) orf_start    - start coordinate of the open reading frame in a
transcript;
iii) orf_end     - end coordinate of the open reading frame in a
transcript.}

\item{gtf}{character string giving the name of output GTF/GFF file with
Ensembl-like annotated experimental transcriptome.}

\item{src}{character string describing the origin of the GTF/GFF file.}

\item{organism}{the proper scientific name of the organism for which
transcriptomic data were generated.}

\item{sqlite}{character string giving the name of output SQLite database
with Ensembl-like annotated experimental transcriptome.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that mean the current working directory.}
}
\value{
GTF/GFF file and local SQLite database with Ensembl-like annotations
of the experimental transcriptome.
}
\description{
This function converts a standard GTF/GFF file with experimental
transcriptome to Ensembl-like annotations.
}
\examples{
anno <- developAnnotations(gtfDir="Files_GTF",
                           gtfFile="test.data_6.gtf",
                           orf="test.data_6.txt",
                           gtf="test.data_6.anno.gtf",
                           src="StringTie",
                           organism="Homo sapiens",
                           sqlite="test.data_6.anno.sqlite",
                           workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev.
}
