% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatingAndMeasuringEffects.r
\name{extractSeqSSs}
\alias{extractSeqSSs}
\title{Extract of splice site sequences from a reference genome}
\usage{
extractSeqSSs(x, genome = "BSgenome.Hsapiens.UCSC.hg38", workDir = NULL)
}
\arguments{
\item{x}{genomic coordinates of 5' or 3' splice sites. It can be the name of
object of class GRanges (as output of extractSSsCoordsEEJs() or
extractSSsCoordsGTF() function) or tab-delimited TXT file with the
following fields:
i) seqnames - name of chromosome or scaffold with prefix "chr";
ii) start   - start coordinate of splice site;
iii) end    - end coordinate of splice site;
iv) strand  - strand information about splice site;
v-...) ...  - (optionally) any metadata.
If TXT file, it should be in work directory.}

\item{genome}{character string giving the name of the BSgenome data package
to be used for reference sequence extraction. Default value is
package BSgenome.Hsapiens.UCSC.hg38.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that mean the current working directory.}
}
\value{
object of class DNAStringSet containing splice site sequences.
}
\description{
Extraction of splice site sequences from the reference genome.
}
\examples{
res <- extractSeqSSs(x="EEJs.Kasumi_1.RR_KD.PMC_BSU.fiveSSs.txt",
                     genome="BSgenome.Hsapiens.UCSC.hg38",
                     workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev, Nadzeya A. Boyeva.
}
