% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparingCallingData.R
\name{filterBamFile}
\alias{filterBamFile}
\title{Filter a BAM file}
\usage{
filterBamFile(
  bamDir = NULL,
  bamFile,
  gr = NULL,
  flag = scanBamFlag(),
  tlen = NULL,
  mapq = NULL,
  index = FALSE,
  workDir = NULL
)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{gr}{character string specifying the name of the tab-delimited TXT file
containing coordinates of genomic location(-s) of interest. The default
value is NULL. If so, this file must contains the following four fields:
i) seqnames - the name of chromosome or scaffold with prefix "chr";
ii) start   - start coordinate of genomic interval of interest;
iii) end    - end coordinate of genomic interval of interest;
iv) strand  - (optionally) strand information about genomic interval
of interest.
TXT file with genomic location(-s) must be in working directory.}

\item{flag}{logical vector used to filter reads based on their flag entry.
This is most easily created with the helper function scanBamFlag() from
package Rsamtools. The function scanBamFlag() recognizes the following
flags: isPaired, isProperPair, isUnmappedQuery, hasUnmappedMate,
isMinusStrand, isMateMinusStrand, isFirstMateRead, isSecondMateRead,
isNotPrimaryRead, isSecondaryAlignment, isNotPassingQualityControls,
isDuplicate and isSupplementaryAlignment. The chosen flag must have
value TRUE to be involved in filtration process and only such flag
should be submitted to the function. Filtering by flag is not performed
by default.}

\item{tlen}{integer value limiting the maximum template length. NULL by
default that mean no any limitations.}

\item{mapq}{integer value limiting the minimum mapping quality. NULL by
default that mean no any limitations.}

\item{index}{logical, FALSE by default. If TRUE, it allows to sort and index
a new filtered BAM file.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
filtered BAM file (optionally, sorted and indexed if index=TRUE)
with postfix "filtered".
}
\description{
Multiparameter BAM file filtering.
}
\examples{
filterBamFile(bamDir="Files_BAM",
              bamFile="SRR21721515.bam",
              gr="genomic_intervals.txt",
              flag=scanBamFlag(isPaired=TRUE, isProperPair=TRUE),
              tlen=300,
              mapq=20,
              index=TRUE,
              workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
