% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteringAndBenchmarkingVariants.r
\name{filterVariants}
\alias{filterVariants}
\title{Filter of single nucleotide variations}
\usage{
filterVariants(
  vcfDir = NULL,
  vcfFile,
  gr = NULL,
  depth = 1,
  score = NULL,
  heterozyg = 0,
  heterozyg1 = 0,
  index = FALSE,
  workDir = NULL
)
}
\arguments{
\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(-s). NULL by default, which means the current
working directory will be used instead of the specified VCF directory.}

\item{vcfFile}{character string giving the name of input VCF file.}

\item{gr}{character string specifying the name of the tab-delimited TXT file
containing coordinates of genomic location(-s) of interest. The default
value is NULL. If so, this file must contains the following four fields:
i) seqnames - the name of chromosome or scaffold with prefix "chr";
ii) start   - start coordinate of genomic interval of interest;
iii) end    - end coordinate of genomic interval of interest;
iv) strand  - (optionally) strand information about genomic interval
of interest.
TXT file with genomic location(-s) must be in working directory.}

\item{depth}{integer value giving the threshold for the minimum sequencing
depth. Default value is 1.}

\item{score}{numeric value that specifies the threshold for the QUAL field.
Default value is NULL.}

\item{heterozyg}{numeric value that specifies the threshold for the total
heterozygosity calculated by the fields MMsum and DP. Default value is 0.}

\item{heterozyg1}{numeric value that specifies the threshold for the
heterozygosity of first alternative allele calculated by the fields MM
and DP. Default value is 0.}

\item{index}{logical agrument, whether to bgzip the output file and generate
a tabix index.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
filtered VCF file with postfix "filtered".
}
\description{
This function performs multi-stage filtering of single
nucleotide variations given in the standard VCF format.
}
\examples{
SNVs_filter <- filterVariants(vcfDir="Files_VCF",
                              vcfFile="example_seq.vcf",
                              gr="genomic_intervals.txt",
                              depth=10,
                              score=5,
                              heterozyg=0.2,
                              heterozyg1=0.2,
                              index=TRUE,
                              workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev.
}
