% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAdapters.r
\name{findAdapters}
\alias{findAdapters}
\title{Identify adapter sequences in short reads}
\usage{
findAdapters(fastqDir = NULL, fastq, n = NULL, adapters, workDir = NULL)
}
\arguments{
\item{fastqDir}{character string specifying the name of the directory
containing the FASTQ file(s). NULL by default that mean the current
working directory.}

\item{fastq}{character string specifying the name of the FASTQ file
containing short reads. Allowed formats are "fastq" or "fastq.gz".}

\item{n}{integer value giving the number of reads loaded into the computer's
RAM. The default value is NULL, at which all reads are loaded into RAM.}

\item{adapters}{character string specifying the name of the tab-delimited
TXT file containing adapter sequences. This file must contain the
following three fields:
i) adapter_id         - adapter ID;
ii) adapter_name      - adapter name;
iii) adapter_sequence - adapter sequence.}

\item{workDir}{character string specifying the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
object of class data frame with binned (by nucleotide positions)
frequency data on the presence of adapter sequences in short reads.
}
\description{
Identification of the adapter sequences in short reads.
}
\examples{
ad <- findAdapters(fastqDir="Files_FASTQ",
                   fastq="2-Galkina-A_S2_L001_R1_001.fastq.gz",
                   n=1e5,
                   adapters="adapters.txt",
                   workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
