% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherCalleR.r
\name{fisherCalleR}
\alias{fisherCalleR}
\title{Call of single nucleotide variations using Fisher’s exact test}
\usage{
fisherCalleR(
  bamDir = NULL,
  bamFile,
  fastaDir = NULL,
  faFile,
  baseq = 20,
  mindepth = 1,
  maxdepth = 1e+06,
  qvalue = 12,
  trim = 0,
  threads = 1,
  workDir = NULL
)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(-s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{fastaDir}{character string giving the name (or path to and name) of
folder for storing of reference genome.}

\item{faFile}{charater string giving the name of file containing reference
genome in format *.FASTA or *.FA.}

\item{baseq}{integer value, minimal "base quality" for each nucleotide in an
alignment. Default value is 20 that means read nucleotides with quality
scores less than 20 will be excluded from analysis.}

\item{mindepth}{integer value giving the minimal coverage of location
containing the variation. Default value is 1.}

\item{maxdepth}{integer value giving the maximal coverage of location
containing the variation. Default value is 1e6. This option is useful
for removing PCR artifacts.}

\item{qvalue}{numeric value giving the q-value cutoff for variations calling
at sequencing depth 50X. Default value is 12. The q-value is calcuated
as -log10(p), where p is a p-value yielded from the Fisher’s exact test.
The function exactSNP() automatically adjusts the q-value cutoff for
each chromosomal location according to its sequencing depth, based on
this cutoff.}

\item{trim}{integer value giving the number of nucleotides trimmed off from
each end of the read. Default value is 0.}

\item{threads}{numeric value giving the number of threads/CPUs used. Default
value is 1.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
file of VCF format is written to the current working directory.
}
\description{
This is a wrapper function that uses the low-level function
exactSNP() from package Rsubread for accurate and efficient SNVs calling.
}
\examples{
res <- fisherCalleR(bamDir="Files_BAM",
                    bamFile="SRR21721515.bam",
                    fastaDir="Files_FASTA",
                    faFile="hg38.fa.gz",
                    baseq=20,
                    mindepth=1,
                    maxdepth=1e6,
                    qvalue=12,
                    trim=0,
                    threads=3,
                    workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev.
}
