% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVCF.r
\name{generateVCF}
\alias{generateVCF}
\title{Generate VCF file based on call table}
\usage{
generateVCF(input, version = "4.0", vcfDir = NULL, outputVCF, workDir = NULL)
}
\arguments{
\item{input}{character string giving the name of tab-delimited TXT file or
R object of class data.frame containing call data. File or object must
contains the following fields:
i) position   - location of single nucleotide variant in format
chrZ:coordinate;
ii) reference - reference variant of nucleotide;
iii) A        - number of redas supporting A in a given position;
iv) C         - number of redas supporting C in a given position;
v) G          - number of redas supporting G in a given position;
vi) T         - number of redas supporting T in a given position;
vii-xi) ...   - one to five columns with statistical test results.
If file, it must be located in working directory.}

\item{version}{character string specifying the version of VCF format.
Default value is "4.0". It is planned that other versions of the format
will be added in the future.}

\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(-s). NULL by default, which means the current
working directory will be used instead of the specified VCF directory.}

\item{outputVCF}{character string giving the name of output VCF file.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
file(-s) of VCF format is/are written to the current working
directory or specified VCF directory.
}
\description{
This function converts call table to standard VCF file.
}
\examples{
generateVCF(input="example_seq.call_table.txt",
            version="4.0",
            vcfDir="Files_VCF",
            outputVCF="example_seq.call_table.identified_SNVs",
            workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev.
}
