% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparingCallingData.R
\name{groupReadsGATK}
\alias{groupReadsGATK}
\title{Group short reads in BAM file}
\usage{
groupReadsGATK(
  bamDir = NULL,
  bamFile,
  postfix = "RG",
  SE = TRUE,
  RGID = "Bam.Recal.RG",
  RGLB = "default.LIBRARY.ID",
  RGPL = "default.ILLUMINA",
  RGPU = "default.PLATFORM",
  RGSM = "default.SAMPLE",
  gatk_path = "gatk",
  workDir = NULL
)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{postfix}{character string that will be added to the output file name
as postfix. "RG" by default.}

\item{SE}{if TRUE (by default), standard error determined by function
AddOrReplaceReadGroup() will be writen in file STDERR.RG.TXT.}

\item{RGID}{RGID value for appropriate BAM field.}

\item{RGLB}{RGLB value for appropriate BAM field.}

\item{RGPL}{RGPL value for appropriate BAM field.}

\item{RGPU}{RGPU value for appropriate BAM field.}

\item{RGSM}{RGSM value for appropriate BAM field.}

\item{gatk_path}{character string giving the path to internal or external
(system) GATK executive script. This path can be obtained by running
Find.Gatk() or settled manually. Default value is "gatk", i.e. function
will uses GATK systemly from $PATH.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.
Both bamDir (when provided) and bamFile must be somewhere inside
directory, chosen as work directory.}
}
\value{
a new BAM file with additional information about read groups and
(optionally) auxillary file with standard error log.
}
\description{
This is a wrapper function that uses the low-level function
AddOrReplaceReadGroup() from package GATK to add read group tags to BAM
file of interest. This stage could be unnecessary, if BAM file already
contains read group tags.
}
\examples{
...
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
