% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparingCallingData.R
\name{indexVcfGATK}
\alias{indexVcfGATK}
\title{Index a VCF file with known variants}
\usage{
indexVcfGATK(
  vcfDir = NULL,
  vcfFile = NULL,
  output_name = NULL,
  gatk_path = "gatk",
  workDir = NULL
)
}
\arguments{
\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(s). NULL by default, which means that the
current working directory will be used instead of the specified VCF
directory.}

\item{vcfFile}{character string giving the name of input VCF file.}

\item{output_name}{character string giving the name of index. NULL by
default, which means that the name of input VCF file will be used.
Important note: to be used correctly by the GATK, the VCF file and its
index must have the same name.}

\item{gatk_path}{character string giving the path to internal or external
(system) GATK executive script. This path can be obtained by running
Find.Gatk() or settled manually. Default value is "gatk", i.e. function
will uses GATK systemly from $PATH.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.
Both vcfDir (when provided) and vcfFile must be somewhere inside
directory, chosen as working directory.}
}
\value{
.tbi file in working directory.
}
\description{
This is a wrapper function that uses the low-level function
IndexFeatureFile() from package GATK to built a .tbi index for VCF file
with known variants.
}
\examples{
...
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
