% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatingAndMeasuringEffects.r
\name{introduceSNVtoAAs}
\alias{introduceSNVtoAAs}
\title{Predict amino acid coding changes for single nucleotide variants}
\usage{
introduceSNVtoAAs(
  vcfDir = NULL,
  vcfFile,
  OrgDb = "org.Hs.eg.db",
  TxDb = "TxDb.Hsapiens.UCSC.hg38.knownGene",
  genome = "BSgenome.Hsapiens.UCSC.hg38",
  proteins,
  workDir = NULL
)
}
\arguments{
\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(-s). NULL by default, which means the current
working directory will be used instead of the specified VCF directory.}

\item{vcfFile}{character string giving the name of input VCF file.}

\item{OrgDb}{character string giving the name of the OrgDb data package with
genome wide annotations. Default value is package org.Hs.eg.db.}

\item{TxDb}{character string giving the name of the TxDb data package with
transcriptome wide annotations. Default value is package
TxDb.Hsapiens.UCSC.hg38.knownGene.}

\item{genome}{character string giving the name of the BSgenome data package
to be used for reference sequence extraction. Default value is
package BSgenome.Hsapiens.UCSC.hg38.}

\item{proteins}{character string giving the name of TXT file in tab-delimited
format with protein coding annotations of Ensembl transcripts. This file
must contains the following IDs: Ensembl transcript IDs, Ensembl gene
IDs, Ensembl protein _IDs, PDB IDs, UniProtKB/Swiss-Prot IDs and
UniProtKB/TrEMBL IDs.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that mean the current working directory.}
}
\value{
object of class data frame with predicted and annotated amino acid
coding changes for single nucleotide variants.
}
\description{
Prediction of amino acid coding changes for single nucleotide
variants.
}
\examples{
vcfFile="test_seq5, fisherCalleR, SNVs.vcf"
proteins="Ensembl_GRCh38.p14_release.114_proteins.txt"
intrSNVtoAAs <- introduceSNVtoAAs(vcfDir="Files_VCF",
                                  vcfFile=vcfFile,
                                  OrgDb="org.Hs.eg.db",
                                  TxDb="TxDb.Hsapiens.UCSC.hg38.knownGene",
                                  genome="BSgenome.Hsapiens.UCSC.hg38",
                                  proteins=proteins,
                                  workDir="D:/Vasily Grinev/SNVs")
}
\author{
Vasily V. Grinev.
}
