% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatingAndMeasuringEffects.r
\name{introduceSNVtoSSs}
\alias{introduceSNVtoSSs}
\title{Introduce single nucleotide variant(-s) to splice site(-s)}
\usage{
introduceSNVtoSSs(
  x,
  vcfDir = NULL,
  vcfFile,
  genome = "BSgenome.Hsapiens.UCSC.hg38",
  workDir = NULL
)
}
\arguments{
\item{x}{genomic coordinates of 5' or 3' splice sites. It can be the name of
object of class GRanges (as output of extractSSsCoordsEEJs() or
extractSSsCoordsGTF() function) or tab-delimited TXT file with the
following fields:
i) seqnames - name of chromosome or scaffold with prefix "chr";
ii) start   - start coordinate of splice site;
iii) end    - end coordinate of splice site;
iv) strand  - strand information about splice site;
v-...) ...  - (optionally) any metadata.
If TXT file, it should be in work directory.}

\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(-s). NULL by default, which means the current
working directory will be used instead of the specified VCF directory.}

\item{vcfFile}{character string giving the name of input VCF file.}

\item{genome}{character string giving the name of the BSgenome data package
to be used for reference sequence extraction. Default value is
package BSgenome.Hsapiens.UCSC.hg38.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that mean the current working directory.}
}
\value{
object of class list containing extended data frame with splice
sites and DNAStringSet with corrected splice site sequences.
}
\description{
Introduction of single nucleotide variant(-s) into the splice
site sequence(-s).
}
\examples{
coords <- "EEJs.Kasumi_1.RR_KD.PMC_BSU.fiveSSs.txt"
vcf <- "test_seq4, fisherCalleR, SNVs.filtered.vcf"
intrSNVtoSSs <- introduceSNVtoSSs(x=coords,
                                  vcfDir="Files_VCF",
                                  vcfFile=vcf,
                                  genome="BSgenome.Hsapiens.UCSC.hg38",
                                  workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev, Nadzeya A. Boyeva.
}
