% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calling.R
\name{mlCalleR}
\alias{mlCalleR}
\title{Call single nucleotide variations using machine learning models}
\usage{
mlCalleR(
  fr_data,
  model = "rpart",
  pvalue = FALSE,
  vcfDir = vcfDir,
  outputVCF = outputVCF,
  workDir = NULL
)
}
\arguments{
\item{model}{character string specifying which machine learning model to use.
Available options: "log_reg", "rand_forest", "rpart", "SVM", "xgboost".
Default is "rpart".}

\item{pvalue}{logical indicating whether to return class probabilities
instead of class labels (default: FALSE). If TRUE, returns probability
of belonging to positive class (1).}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
VCF file with detected single nucleotide variations.
}
\description{
This function uses pre-trained machine learning models to
identify single nucleotide variations from frequency table data.
}
\examples{
# Example usage
mlSNVs <- mlCalleR(model = "rpart", pvalue = FALSE, workDir = "D:/Vasily Grinev")
}
\author{
Dzianis D. Sarnatski
}
