% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFASTQ.r
\name{readFASTQ}
\alias{readFASTQ}
\title{Read records from FASTQ-formatted file}
\usage{
readFASTQ(fastqDir = NULL, fastq, n = NULL, workDir = NULL)
}
\arguments{
\item{fastqDir}{character string specifying the name of the directory
containing the FASTQ file(s). NULL by default that mean the current
working directory.}

\item{fastq}{character string specifying the name of the FASTQ file
containing short reads. Allowed formats are "fastq" or "fastq.gz".}

\item{n}{integer value giving the number of reads loaded into the computer's
RAM. The default value is NULL, at which all reads are loaded into RAM.}

\item{workDir}{character string specifying the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
object of class list containing DNAStringSet object of read
sequences and PhredQuality object of read quality scores.
}
\description{
Reading the records from FASTQ-formatted file.
}
\examples{
fq <- readFASTQ(fastqDir="Files_FASTQ",
                fastq="example_seq.read1.fastq.gz",
                n=1e5,
                workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
