% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsAligment.R
\name{reportQAlResults}
\alias{reportQAlResults}
\title{Collect and present alignment quality assessment results}
\usage{
reportQAlResults(x, output = NULL, workDir = NULL)
}
\arguments{
\item{x}{a character string specifying the name of object developed by the
function assessQAlignedReads().}

\item{output}{a character string that specifying a postfix for naming all
output report files. The default value is NULL, at which the name of
sample will be used for naming all output files.}

\item{workDir}{character string specifying the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
a series of text and graphic files.
}
\description{
A high-level function for consolidating and presenting the
results of evaluation of the alignment quality metrics for aligned
DAN-Seq and/or RNA-Seq short reads.
}
\examples{
report <- reportQAlResults(x=qal,
                           output=NULL,
                           workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
