% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawReadsQC.R
\name{reportQAResults}
\alias{reportQAResults}
\title{Collect and present the results of the quality assessment of short reads}
\usage{
reportQAResults(x, output = NULL, workDir = NULL)
}
\arguments{
\item{x}{character string specifying the name of object developed by the
function assessQRawReads().}

\item{output}{character string that specifying a postfix for naming all
output report files. The default value is NULL, at which the name of
sample will be used for naming all output files.}

\item{workDir}{character string specifying the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
a series of text and graphic files.
}
\description{
A high-level function for consolidating and presenting the
results of quality assessment of short reads generated by the function
assessQRawReads().
}
\examples{
report <- reportQAResults(x=qa,
                          output=NULL,
                          workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
