% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snp_counting_cpp_parallel}
\alias{snp_counting_cpp_parallel}
\title{Parallel SNP Counting in a DataFrame}
\usage{
snp_counting_cpp_parallel(df_d)
}
\arguments{
\item{df_d}{A data frame containing the following columns:
\itemize{
\item \code{reference}: A character vector with the reference alleles ("A", "C", "G", "T").
\item \code{A}, \code{C}, \code{G}, \code{T}: Integer vectors representing the counts of each nucleotide at each position.
}}
}
\value{
A numeric vector containing the ratio of the reference allele count to the total allele count for each position.
If the total count is zero, the result will be \code{NA}.
}
\description{
This function calculates the reference allele count ratio for each position in the provided genomic data.
It uses parallel processing for faster execution by distributing the computation across multiple cores.
The ratio is computed as the reference allele count divided by the total allele count at each position.
The function uses the Rcpp and RcppParallel libraries to efficiently handle the calculations in parallel.
}
\details{
The function performs the following steps:
\enumerate{
\item For each position, the total number of nucleotide counts (\code{A[i]}, \code{C[i]}, \code{G[i]}, \code{T[i]}) is computed.
\item The reference allele count is determined based on the \code{reference} column, comparing it to the nucleotide counts.
\item The ratio of the reference allele count to the total count is calculated for each position.
\item The calculation is performed in parallel across multiple cores to speed up the process for large datasets.
}
}
\examples{
df <- data.frame(reference = c("A", "C", "G", "T"),
                 A = c(10, 0, 5, 0),
                 C = c(2, 10, 0, 0),
                 G = c(0, 0, 10, 5),
                 T = c(0, 0, 0, 10))

result <- snp_counting_cpp_parallel(df)
print(result)

}
\author{
Dzianis D. Sarnatski, Mikalai M. Yatskou
}
