% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortBamFile.r
\name{sortBamFile}
\alias{sortBamFile}
\title{Sort and index BAM file(-s)}
\usage{
sortBamFile(bamDir, bamFile, byQname, workDir = NULL)
}
\arguments{
\item{bamDir}{character vector giving the path to and name of directory
with BAM file(s).}

\item{bamFile}{character vector giving the name(-s) of BAM file(-s) to be
sorted and indexed.}

\item{byQname}{logical. It indicats whether the sorted destination file
should be sorted by query-name (TRUE) or by mapping position (FALSE).}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
sorted and indexed BAM file(-s).
}
\description{
This is a wrapper function that uses the low-level function
sortBam() from package Rsamtools to sort and index BAM file(-s).
}
\examples{
sortBamFile(bamDir="Files_BAM",
            bamFile="SRR21721515.bam",
            byQname=TRUE,
            workDir="D:/Vasily Grinev")
}
\author{
Vasily V. Grinev
}
