% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatingAndMeasuringEffects.r
\name{annotateVariants}
\alias{annotateVariants}
\title{Annotate of single nucleotide variations}
\usage{
annotateVariants(
  vcfDir = NULL,
  vcfFile,
  ref_genes,
  output = NULL,
  workDir = NULL
)
}
\arguments{
\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(-s). NULL by default, which means the current
working directory will be used instead of the specified VCF directory.}

\item{vcfFile}{character string giving the name of input VCF file.}

\item{ref_genes}{character string giving the name of TXT file in
tab-delimited format with basic annotations of reference genes. This
file must contains the following fields:
i) gene_id        - (optionally) gene ID (for example, Ensembl based);
ii) gene_name     - name of gene (for example, HUGO based);
iii) gene_biotype - biotype of gene;
iv) seqnames      - name of chromosome or scaffold with prefix "chr";
v) start          - start coordinate of gene;
vi) end           - end coordinate of gene;
vii) strand       - (optionally) strand information about gene.}

\item{output}{character string giving the name (without extension) of TXT
file in tab-delimited format for storing annotated single nucleotide
variations. NULL by default, which means that the name will be generated
automatically based on the name of the input VCF file.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
an object of class data.frame and new TXT file in tab-delimited
format containing annotated single nucleotide variations.
}
\description{
This function performs a multivariate annotation of the single
nucleotide variations given in the standard VCF format.
}
\examples{
genes <- "Ensembl release 114, GRCh38.p14, annotations of genes.txt"
annoSNVs <- annotateVariants(vcfDir="Files_VCF",
                             vcfFile="example_seq.vcf",
                             ref_genes,
                             output="example_seq, annotated SNVs",
                             workDir="D:/Vasily Grinev")
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
