% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatkCalleR.R
\name{gatkCalleR}
\alias{gatkCalleR}
\title{Call of single nucleotide variations using haplotype caller GATK}
\usage{
gatkCalleR(
  bamDir = NULL,
  bamFile,
  postfix = NA,
  fastaDir = NULL,
  faFile,
  intervals = NULL,
  pcr_model,
  soft_clip = TRUE,
  ERC = "NONE",
  cores = "MAX",
  SE = TRUE,
  gatk_path = NULL,
  workDir = NULL
)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{postfix}{character string that will be added to the output VCF file
name as postfix. NA by default at which the postfix "GATK based SNVs"
will be used.}

\item{fastaDir}{character string giving the name (or path to and name) of
folder for storing of fai-indexed reference genome.}

\item{faFile}{charater string giving the name of file containing reference
genome in format *.FASTA or *.FA.}

\item{intervals}{character string specifying the name of the tab-delimited
TXT file containing coordinates of genomic location(-s) of interest. The
default value is NULL. If so, this file must contains the following four
fields:
i) seqnames - the name of chromosome or scaffold with prefix "chr";
ii) start   - start coordinate of genomic interval of interest;
iii) end    - end coordinate of genomic interval of interest;
iv) strand  - (optionally) strand information about genomic interval
of interest.
TXT file with genomic location(-s) must be in working directory.}

\item{pcr_model}{GATK argument --pcr-indel-model. Possible values: NONE,
AGRESSIVE, HOSTILE, CONSERVATIVE. See HaplotypeCaller
at https://gatk.broadinstitute.org for further details.}

\item{soft_clip}{GATK argument --dont-use-soft-clipped. TRUE by default.}

\item{ERC}{GATK ERC mode. "NONE" by default. Possible values: NONE, GVCF or
BP_RESOLUTION. See HaplotypeCaller at https://gatk.broadinstitute.org
for further details.}

\item{cores}{number of CPU cores to use for Hidden Markov Models. If "MAX"
(by default), all available cores will be detected and used.}

\item{SE}{if TRUE (by default), standart error determined by GATK will be
writen in file STDERR.HC.TXT.}

\item{gatk_path}{character string giving the path to internal or external
(system) GATK executive script. This path can be obtained by running
Find.Gatk() or settled manually. Default value is "gatk", i.e. function
will uses GATK systemly from $PATH.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
a new VCF file with detected SNVs.
}
\description{
This is a wrapper function to run GATK within the R environment
for accurate and efficient SNVs calling.
}
\details{
GATK performs SNVs and short indels calling by local de novo
assembly of mapped reads
}
\examples{
...
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
