% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGenomeGATK.r
\name{indexGenomeGATK}
\alias{indexGenomeGATK}
\title{Index a reference genome}
\usage{
indexGenomeGATK(
  fastaDir = NULL,
  faFile,
  fai = TRUE,
  dict = TRUE,
  gatk_path = "gatk",
  workDir = NULL
)
}
\arguments{
\item{fastaDir}{character string giving the name (or path to and name) of
folder for storing of reference genome.}

\item{faFile}{charater string giving the name of file containing reference
genome in format *.FASTA or *.FA.}

\item{fai}{logical. If TRUE (by default), .fai index of FASTA file will be
created in working diectory.}

\item{dict}{logical. If TRUE (by default), .dict index of FASTA file will be
created in working diectory.}

\item{gatk_path}{character string giving the path to internal or external
(system) GATK executive script. This path can be obtained by running
Find.Gatk() or settled manualy. Default value is "gatk", i.e. function
will uses GATK systemly from $PATH.}

\item{workDir}{character string giving the path to and name of working
directory. NULL by default that means the current working directory.}
}
\value{
index files in working directory.
}
\description{
This is a wrapper function that uses the low-level function
CreateSequenceDictionary() from package GATK to built .dict and/or .fai
indices for reference genome of interest.
}
\examples{
...
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
