% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparingCallingData.R
\name{markDuplicatesGATK}
\alias{markDuplicatesGATK}
\title{Detect and mark duplicated reads}
\usage{
markDuplicatesGATK(
  bamDir = NULL,
  bamFile,
  postfix = "markedDuplicates",
  removeDupls = "no",
  SE = TRUE,
  gatk_path = "gatk",
  workDir = NULL
)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{postfix}{character string that will be added to the output file name
as postfix. "markedDuplicates" by default.}

\item{removeDupls}{a way to deal with detected duplicated reads. With "no"
all duplicated reads will be kept in output BAM file, with "yes" all
duplicated reads will be removed, and with "seq" only optical and other
sequencing duplicates will be removed.}

\item{SE}{if TRUE (by default), standart error determined by function
MarkDuplicates() will be writen in file STDERR.MARKDUP.TXT.}

\item{gatk_path}{character string giving the path to internal or external
(system) GATK executive script. This path can be obtained by running
Find.Gatk() or settled manually. Default value is "gatk", i.e. function
will uses GATK systemly from $PATH.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
a new BAM file with processed duplicated reads, a file with metrics
of input BAM file and (optionally) a file with standart error log.
}
\description{
This is a wrapper function that uses the low-level function
MarkDuplicates() from package GATK to find out and mark duplicated reads
in BAM file(-s) of interest.
}
\examples{
...
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
