% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrateBamGATK.r
\name{recalibrateBamGATK}
\alias{recalibrateBamGATK}
\title{Recalibrate a BAM file}
\usage{
recalibrateBamGATK(
  bamDir = NULL,
  bamFile,
  vcfDir = NULL,
  vcfFile,
  fastaDir = NULL,
  faFile,
  createBQSR = TRUE,
  recalibrate = TRUE,
  nameBQSR,
  postfix = "Recalibrated",
  SE = TRUE,
  gatk_path = "gatk",
  workDir = NULL
)
}
\arguments{
\item{bamDir}{character string giving the name (or path to and name) of
directory with BAM file(s). NULL by default, which means the current
working directory will be used instead of the specified BAM directory.}

\item{bamFile}{character string giving the name of input BAM file.}

\item{vcfDir}{character string giving the name (or path to and name) of
directory with VCF file(s). NULL by default, which means that the
current working directory will be used instead of the specified VCF
directory.}

\item{vcfFile}{character string giving the name of input VCF file.}

\item{fastaDir}{character string giving the name (or path to and name) of
folder for storing of reference genome.}

\item{faFile}{charater string giving the name of file containing reference
genome in format *.FASTA or *.FA.}

\item{createBQSR}{logical, TRUE by default. It allows to create a new BQSR
table with provided BAM file, known variation sites and reference genome.}

\item{recalibrate}{logical, TRUE by default. It allows to run a recalibration
process. For this purpose, the BQSR table provided trough nameBQSR or
created by function itself if createBQSR=TRUE, will be used.}

\item{nameBQSR}{character string giving the name of pre-computed BQSR table.
NULL by default.}

\item{postfix}{character string that will be added to the output file name
as postfix. "Recalibrated" by default.}

\item{SE}{if TRUE (by default), standart error determined by function
MarkDuplicates() will be writen in file STDERR.RECALIBRATION.TXT.}

\item{gatk_path}{character string giving the path to internal or external
(system) GATK executive script. This path can be obtained by running
Find.Gatk() or settled manually. Default value is "gatk", i.e. function
will uses GATK systemly from $PATH.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
if createBQSR=TRUE, a new BQSR table in working directory;
if recalibrate=TRUE, a new BAM file with recalibrated sequencing quality
scores of bases in working directory; if SE=TRUE, a file with standart
error log in working directory.
}
\description{
This is a wrapper function that uses the low-level functions
BaseRecalibrator() and ApplyBQSR() from package GATK to recalibrate
sequencing quality scores of bases in BAM file of interest.
}
\examples{
...
}
\author{
Ilia M. Ilyushonak, Vasily V. Grinev.
}
